wip: some basics to make it sort of use MLOS

+++ external/ExternalIntegrationExample/ExternalIntegrationExample.SettingsRegistry/CodeGen/SmartComponentExampleConfig.cs
@@ -32,6 +32,6 @@ namespace ExternalIntegrationExample
         /// The size for the smart component example.
         /// </summary>
         [ScalarSetting]
-        internal int Size;
+        internal long Size;
     }
 }
